+++ docs/misc/changelog.rst
@@ -41,6 +41,7 @@ Others:
 - Fixed ``stable_baselines3/ppo/*.py`` type hints
 - Fixed ``stable_baselines3/sac/*.py`` type hints
 - Fixed ``stable_baselines3/td3/*.py`` type hints
+- Fixed ``stable_baselines3/common/base_class.py`` type hints
 
 Documentation:
 ^^^^^^^^^^^^^^